+++ importSOCAT.py
@@ -15,7 +15,7 @@ flagaccuracy = {"A": 2.0, "B": 2.0, "C": 5.0, "D": 5.0, "E": 10.0}
 
 # Read from local files (synthesis + flagE
 if (datafrom == 'local'):
-    filespath=filespathlocal
+    filespath = filespathlocal
     SOCATmetacolheadertextshort = 'Expocode\tversion\tDataset'
     SOCATcolheadertextshort = 'Expocode\tversion\tSource_DOI\tQC_Flag'
 
@@ -42,8 +42,8 @@ if (datafrom == 'local'):
             metaline = f.readline()
             endmetaheaderlines = endmetaheaderlines + 1  # Where metadata lines end
         # Create metadata dataframe
-        metainfoAD = pd.read_csv(filespath + file, sep='\t', skiprows=metaheaderlines,
-                                 nrows=endmetaheaderlines - metaheaderlines - 1)
+        metainfoAD = pd.read_csv(filespath + file, sep = '\t', skiprows = metaheaderlines,
+                                 nrows = endmetaheaderlines - metaheaderlines - 1)
         # Find where data columns start
         headerlines = headerlines + endmetaheaderlines
         while SOCATcolheadertextshort not in line:
@@ -56,13 +56,13 @@ if (datafrom == 'local'):
         start_time = time.time()  # Time the script
         ddtype = {0: str, 2: str}  # add type str to columns 0 and 2
         # Read the SOCAT file into a pandas dataframe
-        tempdf1 = pd.read_csv(filespath + file, sep=separator, skiprows=headerlines, dtype=ddtype)
+        tempdf1 = pd.read_csv(filespath + file, sep = separator, skiprows = headerlines, dtype = ddtype)
         print("--- %s seconds ---" % (time.time() - start_time))
         print(file + " data frame has " + str(len(tempdf1)) + " lines")
 
         # Give all the same variable names
         tempdf1.rename(
-            columns={'Expocode': vardict['id'], 'Source_DOI': vardict['doi'],
+            columns = {'Expocode': vardict['id'], 'Source_DOI': vardict['doi'],
                      'latitude [dec.deg.N]': vardict['lat'],
                      'longitude [dec.deg.E]': vardict['lon'],
                      'sample_depth [m]': vardict['dep'], 'SST [deg.C]': vardict['temp'], 'sal': vardict['sal'],
@@ -72,20 +72,20 @@ if (datafrom == 'local'):
         tempdtframe = pd.DataFrame(
             {'year': tempdf1['yr'], 'month': tempdf1['mon'], 'day': tempdf1['day'],
              'hour': tempdf1['hh'], 'minute': tempdf1['mm'], 'seconds': tempdf1['ss']})
-        tempdf1['DATEVECTOR1'] = pd.to_datetime(tempdtframe,utc=True)
+        tempdf1['DATEVECTOR1'] = pd.to_datetime(tempdtframe,utc = True)
 
         # Transform longitude to +-180
-        tempdf1.loc[tempdf1[vardict['lon']] > 180, vardict['lon']] = tempdf1[vardict['lon']] -360
+        tempdf1.loc[tempdf1[vardict['lon']] > 180, vardict['lon']] = tempdf1[vardict['lon']] - 360
 
         # Subset the dataset HERE (cruise flag assignment takes A LONG TIME).
         if subset :
             tempdf1 = tempdf1[(tempdf1['DATEVECTOR1'] >= pd.to_datetime(mindate)) &
                             (tempdf1['DATEVECTOR1'] <= pd.to_datetime(maxdate)) &
-                            (tempdf1[vardict['lat']]>= minlat) &
-                            (tempdf1[vardict['lat']]<= maxlat) &
-                            (tempdf1[vardict['lon']]>= minlon) &
-                            (tempdf1[vardict['lon']]<= maxlon)].copy()
-            tempdf1.reset_index(drop=True, inplace=True)
+                            (tempdf1[vardict['lat']] >= minlat) &
+                            (tempdf1[vardict['lat']] <= maxlat) &
+                            (tempdf1[vardict['lon']] >= minlon) &
+                            (tempdf1[vardict['lon']] <= maxlon)].copy()
+            tempdf1.reset_index(drop = True, inplace = True)
             print(len(tempdf1))
 
         # Cruise flags / accuracies
@@ -121,8 +121,8 @@ if (datafrom == 'local'):
 # If "remote"
 elif (datafrom == 'remote'):
     e = ERDDAP(
-        server='https://data.pmel.noaa.gov/socat/erddap',
-        protocol='tabledap',
+        server = 'https://data.pmel.noaa.gov/socat/erddap',
+        protocol = 'tabledap',
     )
 
     e.response = 'csv'
@@ -140,22 +140,23 @@ elif (datafrom == 'remote'):
         'WOCE_CO2_water=': "2" #synthesis file only has good data (keep questionable/bad?)
         # 'fCO2_water_sst_100humidity_uatm=~':"float('nan')" # Have yet to figure out how to set the nan filter
     }
-    e.variables = ['expocode','time','latitude','longitude','depth','sal','temp',
-                   'fCO2_recommended','qc_flag','WOCE_CO2_water','socat_doi']
-    tempdf = e.to_pandas(dtype={10: str, 8: str, 0: str})
+    e.variables = ['expocode', 'time', 'latitude', 'longitude', 'depth', 'sal',
+                   'temp', 'fCO2_recommended', 'qc_flag', 'WOCE_CO2_water',
+                   'socat_doi']
+    tempdf = e.to_pandas(dtype = {10: str, 8: str, 0: str})
 
     # Retain only valid fco2 values (can't figure out how to do it in erdappy constrains yet)
-    tempdf=tempdf.dropna(subset=['fCO2_recommended (uatm)']).copy()
-    tempdf.reset_index(drop=True, inplace=True)
+    tempdf = tempdf.dropna(subset = ['fCO2_recommended (uatm)']).copy()
+    tempdf.reset_index(drop = True, inplace = True)
 
     # Rename columns
     tempdf.rename(
-        columns={'expocode': vardict['id'], 'socat_doi':vardict['doi'],
+        columns = {'expocode': vardict['id'], 'socat_doi':vardict['doi'],
                  'latitude (degrees_north)': vardict['lat'], 'longitude (degrees_east)': vardict['lon'],
                  'depth (m)': vardict['dep'], 'temp (degrees C)': vardict['temp'], 'sal (PSU)': vardict['sal'],
                  'fCO2_recommended (uatm)': vardict['fco2w'], 'WOCE_CO2_water': vardict['fco2wf'],
                  'qc_flag':'Cruise_flag'},
-        inplace=True)
+        inplace = True)
 
     # Create python date object
     tempdf['DATEVECTOR1'] = pd.to_datetime(tempdf['time (UTC)'])
@@ -170,7 +171,7 @@ tempdf[vardict['fco2wac']] = 0.0
 for key in flagaccuracy:
     tempdf[vardict['fco2wac']].values[tempdf['Cruise_flag'] == key] = flagaccuracy[key]
 # Flag fco2 as measured
-tempdf[vardict['fco2wc']]=0
+tempdf[vardict['fco2wc']] = 0
 
 # Estimate alkalinity from salinity, and then, estimate ph and dic
 
@@ -181,7 +182,7 @@ tempdf.loc[tempdf[vardict['doi']].isna(), vardict['doi']] = socatdoi
 tempdf['SOURCE'] = source
 
 # Rename and reset indices
-printdf=tempdf
+printdf = tempdf
 printdf.reset_index(drop=True, inplace=True)
 
 print('SOCAT frame size is ')
